requirements.txt
pandas
numpy
scikit-learn
matplotlib
seaborn
django
djangorestframework
joblib
